handel some chalange

+++ app/app/models/vertex.py
@@ -8,12 +8,12 @@ class Vertex(Base):
     id: Mapped[int] = mapped_column(
         Integer, primary_key=True, autoincrement=True, index=True
     )
-    endx: Mapped[int] = mapped_column(Float, nullable=True)
-    endy: Mapped[int] = mapped_column(Float, nullable=True)
-    startx: Mapped[int] = mapped_column(Float, nullable=True)
-    starty: Mapped[int] = mapped_column(Float, nullable=True)
-    pox: Mapped[int] = mapped_column(Float, nullable=True)
-    poy: Mapped[int] = mapped_column(Float, nullable=True)
+    endx: Mapped[float] = mapped_column(Float, nullable=True)
+    endy: Mapped[float] = mapped_column(Float, nullable=True)
+    startx: Mapped[float] = mapped_column(Float, nullable=True)
+    starty: Mapped[float] = mapped_column(Float, nullable=True)
+    pox: Mapped[float] = mapped_column(Float, nullable=True)
+    poy: Mapped[float] = mapped_column(Float, nullable=True)
 
     source_edges: Mapped[list["Edge"]] = relationship(
         "Edge", foreign_keys="[Edge.source_vertex_id]", back_populates="source_vertex"
